Set task completed when pressing the icon
Now you can press the icon to set the task completed without passing from the timer.
There's a bug that each time the recycler view updates after setting a task done for a brief time the task for the current day is loaded before the actual day selected. It can be annoying so TO FIX.